Summary: save checkpoint in hex format
This is to avoid using outside utility like elf2hex or bin2hex. Hex format is
used when loading binaries to RTL. This commit adds line option "--save_format".
The default value is 0, which dumps values in binary and if set to 1 the check-
point is dumped to .hex files.

Second, checkpoint files were added to .gitignore.

Lastly, run/checkpoints directory was added. Also, the sample qsort binary with
the corresponding .cfg was added.