Summary: don't compare insts specified in cfg
Note: this change might be reverted if I come up with something
cleaner.

We step dromajo at commit. However, some implementations doesn't
commit some control flow instructions. For example, Ariane doesn't
commit ebreaks and ecalls.

This change is workaround that skips comparisons of specific
instructions specified in cfg. In dromajo_step() method we
check if the dromajo is trying to commit one of those instructions.
In case of a match it returns 0x2 to the DPI wrapper. The wrapper,
in turn, calls the step function one more time. This allows the
ebreaks and ecalls to be skipped and perform comparison of the
following instructions.