Summary: embedding fuzzers into Dromajo
This implements the fuzzer class. The instance of fuzzer class is
the part of RISCVState struct. The idea is to keep all different
fuzzers in this class. The state of the fuzzer will be gotten from
the RTL through DPI. For example, the I implemented congestor in
Ariane.

  logic congest_icache, congest_dcache;
  always_ff @(posedge clk_i or negedge rst_ni) begin
    if (~rst_ni) begin
      congest_icache <= 1'b0;
      congest_dcache <= 1'b0;
    end else begin
      congest_icache <= logic'(dromajo_get_congestor(0));
      congest_dcache <= logic'(dromajo_get_congestor(1));
    end
  end

where dromajo_get_congestor(fuzzer_id) is the wrapper to Dromajo
API as it is usually done.

The fuzzer will be extended with table mutator, random address
generetor, etc.