Summary: implementation of table mutators
Table mutators allow RTL memories to be fuzzed. For example, the
tables of branch predictors can be freely mutated at any given
moment as they shouldn't not have any functional effect. Another
example are the cache or TLB entries. The entries that are invalid
can be randomized runtime.

The table mutator class should implement access memory method. This
method should emulate the memory access of the implementation and
must be called with DPI.